user auth. new screen for auth, sign up, sign in
